requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
Flask-CORS==3.0.10
fpdf==1.7.2
psycopg2-binary==2.9.3
python-dotenv==0.19.0
gunicorn==20.1.0
supabase
reportlab==4.0.4
Pillow==10.0.0
python-dateutil==2.8.2
